pyproject.toml
[project]
name = "shadow-farmer"
version = "0.1.0"
description = ""
authors = [
    {name = "Frank Oge",email = "ogekcfrankie@gmail.com"}
]
readme = "README.md"
requires-python = "^3.12"
dependencies = [
    "fastapi (>=0.128.0,<0.129.0)",
    "uvicorn (>=0.40.0,<0.41.0)",
    "sqlalchemy (>=2.0.45,<3.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "pydantic (>=2.12.5,<3.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "celery (>=5.6.1,<6.0.0)",
    "redis (>=7.1.0,<8.0.0)",
    "web3 (>=7.14.0,<8.0.0)",
    "solana (>=0.36.10,<0.37.0)",
    "playwright (>=1.57.0,<2.0.0)",
    "cryptography (>=46.0.3,<47.0.0)",
    "fernet (>=1.0.1,<2.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "aiohttp (>=3.13.2,<4.0.0)",
    "walletconnect (>=0.1.0a0,<0.2.0)",
    "pyarmor (>=9.2.3,<10.0.0)",
    "playwright-stealth (>=2.0.0,<3.0.0)"
]


[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=9.0.2,<10.0.0)",
    "pytest-asyncio (>=1.3.0,<2.0.0)"
]
